pyproject.toml
[tool.poetry]
name = "fontmapster"
version = "0.0.1"
description = "动态字体映射工具"
authors = ["ayuge <ayugesheng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
fonttools = "4.49.0"
numpy = [
    { version = "~1.24.3", python = "<3.9" },
    { version = "~1.26.0", python = ">=3.9,<3.13" },
]
pillow = "^10.2.0"
brotli = "^1.1.0"
rapidocr-onnxruntime = { version = "^1.3.15", optional = true }
cnocr = { version = "^2.3.0.1", optional = true }

[tool.poetry.group.dev.dependencies]
black = "24.2.0"
isort = "5.13.2"
coverage = "7.2.2"
pytest = "7.1.3"
tox = "4.14.1"

[tool.poetry.extras]
all = ["rapidocr-onnxruntime", "cnocr"]
rapidocr = ["rapidocr-onnxruntime"]
cnocr = ["cnocr"]

[tool.pytest.ini_options]
xfail_strict = true
python_files = "test_*.py __init__.py"
python_classes = "Test*"
testpaths = [
    "tests",
]
norecursedirs = ["dist", "build", "docs", "examples"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:typing.re is deprecated",
    "ignore:typing.io is deprecated"
]

[tool.coverage.run]
branch = true
source = ["fontmapster"]
parallel = true
omit = [
    "tests/fonts/*",
    "tests/**/__init__.py",
    "fontmapster/**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == '__main__':",
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "__repr__",
    "__str__",
]
precision = 2
fail_under = 30

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
